factor out tests
